Export MasterMind logic in its own module
